actions/requirements.txt
rasa-sdk==3.6.2
psycopg2-binary
requests